+++ source/B2B.Transactions/Infrastructure/Authentication/Errors/AuthenticationError.cs
@@ -1,0 +1,12 @@
+﻿namespace B2B.Transactions.Infrastructure.Authentication.Errors
+{
+    public abstract class AuthenticationError
+    {
+        protected AuthenticationError(string message)
+        {
+            Message = message;
+        }
+
+        protected string Message { get; set; }
+    }
+}
