+++ source/B2B.Transactions/Infrastructure/Authentication/Errors/NoAuthenticationHeaderSet.cs
@@ -1,0 +1,10 @@
+﻿namespace B2B.Transactions.Infrastructure.Authentication.Errors
+{
+    public class NoAuthenticationHeaderSet : AuthenticationError
+    {
+        public NoAuthenticationHeaderSet()
+        : base("No authorization header is set.")
+        {
+        }
+    }
+}
