+++ source/B2B.Transactions/Infrastructure/Authentication/Errors/TokenValidationFailed.cs
@@ -1,0 +1,10 @@
+﻿namespace B2B.Transactions.Infrastructure.Authentication.Errors
+{
+    public class TokenValidationFailed : AuthenticationError
+    {
+        public TokenValidationFailed(string message)
+         : base(message)
+        {
+        }
+    }
+}
